Part 9.3
Fixed transporter/spawnpoint glitch
Fixed health bar not showing up
Known glitches:
NPCs rendering behind buildings when they shouldnt
Projectiles do not show up (Same as healthbar problem)
Infinite chest opening
Typo in chest message
Shooting/Swinging animations not working
Demons/Knights moving really fast when attacking

+++ src/game/entities/particles/BlackHole.java
@@ -48,7 +48,7 @@ public class BlackHole extends Particle {
 	 * @param y
 	 *            the y coord A THE CENTER
 	 */
-	public BlackHole(Level level, double x, double y) {
+	public BlackHole(Level level, int x, int y) {
 		super(level, x - SIZE / 2, y - SIZE / 2, 0, new int[] { 0xFF000000, 0xFF000000, 0xFF000000 });
 
 		setSpriteSheet(SpriteSheet.explosions);
